+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@
 * #617 ECharts: "map" series
 * #619 CsvLoader: own CSV parser implementation
 * #621 ECharts: VisualMap.splitNumber
+* #622 ECharts: VisualMap dimension mapping
 
 ## Release 2.0.0-M6